+++ app/access-control/controllers/ldap-controller.html
@@ -7,7 +7,7 @@
     <p>Configure LDAP settings and manage role groups.</p>
   </div>
   <div class="row column">
-    <h2 class="page-subhead">
+    <h2 class="subhead">
       Settings
     </h2>
   </div>
@@ -136,7 +136,7 @@
   </div>
 </div>
 <div class="ldap-groups row column">
-  <h2 class="small-12 page-subhead">
+  <h2 class="small-12 subhead">
     Role groups
   </h2>
   <div class="row column">
