--- app/configuration/styles/certificate.scss
@@ -1,240 +1,0 @@
-.certificate__table {
-  border-left: 1px solid $border-color-01;
-  border-right: 1px solid $border-color-01;
-  margin-top: 0.5em;
-  .table__row-header {
-    width: 100%;
-    border-bottom: 1px solid $border-color-01;
-    background-color: $primary-dark;
-    color: $primary-light;
-    font-weight: 700;
-  }
-  .table__row-value {
-    width: 100%;
-    border-bottom: 1px solid $border-color-01;
-  }
-  .certificate__type-header {
-    @include mediaQuery(small) {
-      width: 20%;
-      background: transparent;
-    }
-    width: 100%;
-    padding: 0.8em;
-    padding-left: 1.5em;
-  }
-  .certificate__issue-header {
-    display: none;
-    padding: 0.8em;
-    @include mediaQuery(small) {
-      width: 20%;
-      display: block;
-    }
-  }
-  .certificate__date-header {
-    padding: 0.8em;
-    @include mediaQuery(small) {
-      width: 12%;
-      display: block;
-    }
-    display: none;
-  }
-  .certificate__status-header {
-    padding: 0.8em;
-    @include mediaQuery(small) {
-      width: 5%;
-      display: block;
-    }
-    display: none;
-  }
-  .certificate__type-cell {
-    width: 100%;
-    padding: 0.8em 0.8em 0.8em 1.5em;
-    word-wrap: break-word;
-    background: $background-02;
-    @include mediaQuery(small) {
-      width: 20%;
-      background: transparent;
-    }
-  }
-  .certificate__issue-cell {
-    padding: 0.8em;
-    word-wrap: break-word;
-    @include mediaQuery(small) {
-      width: 20%;
-    }
-    width: 70%;
-  }
-  .certificate__date-cell {
-    width: 70%;
-    padding: 0.8em;
-    word-wrap: break-word;
-    @include mediaQuery(small) {
-      width: 12%;
-    }
-  }
-  .certificate__status-cell {
-    padding: 0.8em;
-    @include mediaQuery(small) {
-      display: block;
-      width: 5%;
-    }
-    display: none;
-  }
-  .certificate__buttons-cell {
-    @include mediaQuery(small) {
-      width: 10%;
-      padding-top: 0.5em;
-    }
-    width: 100%;
-    text-align: right;
-    padding: 0 1.5em 0 0;
-  }
-  .certificate__title-inline {
-    @include mediaQuery(small) {
-      display: none;
-    }
-    width: 30%;
-    display: block;
-    padding: 0.8em 0.8em 0.8em 1.5em;
-  }
-  .upload__certificate {
-    border-top: 1px solid $border-color-01;
-    width: 100%;
-    background: $background-02;
-    padding: 0.8em;
-  }
-}
-.certificate__upload-chooser {
-  background: $background-02;
-}
-
-
-.certificate__close-modal {
-  float: right;
-  position: relative;
-  bottom: 1rem;
-  left: 2rem;
-}
-
-.add__certificate__modal {
-  select {
-    margin-bottom: 0;
-  }
-  .file__upload {
-    margin-bottom: 2em;
-  }
-  .select__new-label {
-    margin-bottom: 1em;
-  }
-  .select__new-button {
-    font-size: 1.2em;
-  }
-  .file__name {
-    background-color: $background-02;
-    padding: 0.5em;
-  }
-}
-
-.add-certificate__section {
-  padding-left: 0;
-}
-
-// Combinator needed to match specificity
-// of default modal settings
-.modal.add-csr__modal {
-  width: 100%;
-  max-height: 100vh;
-  overflow-y: auto;
-  z-index: 1001;
-}
-
-.add-csr__section:first-of-type {
-  padding-left: 0;
-}
-
-.add-csr__section:last-of-type {
-  margin-top: 2rem;
-  padding-right: 0;
-
-  @media (min-width: 640px) {
-    margin-top: 0;
-  }
-}
-
-.add-csr__section-title {
-  margin-bottom: 1rem;
-  font-weight: 700;
-}
-
-.add-csr__section--border {
-  @media (min-width: 640px) {
-    padding-left: 2rem;
-    border-left: 1px solid $base-02--04;
-  }
-}
-
-.add-csr__label {
-  white-space: nowrap;
-  display: inline-block;
-}
-
-.add-csr__text-helper {
-  color: $base-02--02;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 1.2;
-  margin-bottom: .5em;
-}
-
-input.add-csr__input,
-select.add-csr__select {
-  width: 100%;
-  margin-bottom: 0;
-  max-height: none;
-  height: auto;
-}
-
-.select.add-csr__select  {
-  line-height: 1.15;
-}
-
-input.add-csr__input-no-validation {
-  margin-bottom: 1.7rem;
-}
-
-.add-csr__additional-alt-names {
-  display: flex;
-}
-
-.add-csr__alt-name-add-btn {
-  padding: 0;
-  @media (min-width: 640px) {
-    margin: 1.9rem 0;
-  }
-}
-
-.add-csr__alt-name-delete-btn {
-  width: 20px;
-  height: 30px;
-  padding: 0;
-
-  icon {
-    margin-right: 0;
-  }
-}
-
-.add-csr-code__header {
-  margin-top: 1em;
-}
-
-.add-csr__container-csr-code {
-  white-space: pre-wrap;
-}
-
-.add-csr__text-download {
-  color: $base-02--08;
-}
-
-select.add-csr__multiselect {
-  height: 14rem;
-}
