+++ app/configuration/styles/index.scss
@@ -3,5 +3,3 @@
 @import "./date-time.scss";
 @import "./firmware.scss";
 @import "./virtual-media.scss";
-@import "./certificate.scss";
-@import "./ldap.scss";
